Nest reservable items underneath item pools
