+++ app/helpers/admin/group_lending_helper.rb
@@ -3,9 +3,5 @@ module Admin
     def item_pool_options
       ItemPool.all.map { |item_pool| [item_pool.name, item_pool.id] }
     end
-
-    def item_pool_options_for_reservable_items
-      ItemPool.uniquely_numbered.map { |item_pool| [item_pool.name, item_pool.id] }
-    end
   end
 end
