+++ app/views/admin/item_pools/_item_pool.html.erb
@@ -4,7 +4,7 @@
   </h4>
   <ul>
     <% item_pool.reservable_items.each do |item| %>
-      <li><%= link_to item.name, [:admin, item] %></li>
+      <li><%= link_to item.name, [:admin, item_pool, item] %></li>
     <% end %>
   </ul>
 </div>
