+++ app/views/admin/item_pools/reservable_items/_form.html.erb
@@ -1,6 +1,5 @@
-<%= form_with(model: reservable_item, url: [:admin, reservable_item], builder: SpectreFormBuilder) do |form| %>
+<%= form_with(model: reservable_item, url: [:admin, reservable_item.item_pool, reservable_item], builder: SpectreFormBuilder) do |form| %>
   <%= form.text_field :name %>
-  <%= form.select :item_pool_id, item_pool_options_for_reservable_items %>
   <%= form.select :status, options_for_select(item_status_options, reservable_item.status || ReservableItem.statuses[:active]), prompt: true, hint: "Only active items can be checked out" %>
 
   <div>
