+++ app/views/admin/item_pools/reservable_items/_reservable_item.html.erb
