+++ app/views/admin/item_pools/reservable_items/index.html.erb
@@ -1,6 +1,6 @@
 <% content_for :header do %>
   <%= index_header "Reservable Items" do %>
-    <%= link_to "New Reservable Item", new_admin_reservable_item_path, class: "btn" %>
+
   <% end %>
 <% end %>
