+++ app/views/admin/item_pools/reservable_items/show.html.erb
