+++ app/views/layouts/admin.html.erb
@@ -32,7 +32,6 @@
 
               <% if ENV["FEATURE_GROUP_LENDING"] == "on" %>
                 <li class="nav-item"><%= link_to "Item Pools", admin_item_pools_path %></li>
-                <li class="nav-item"><%= link_to "Reservable Items", admin_reservable_items_path %></li>
                 <li class="nav-item"><%= link_to "Reservations", admin_reservations_path %></li>
                 <li class="nav-item"><%= link_to "Pickups", admin_pickups_path %></li>
               <% end %>
@@ -129,7 +128,6 @@
                 <% if ENV["FEATURE_GROUP_LENDING"] == "on" %>
                   <li class="menu-item"><div class="divider"></div></li>
                   <li class="menu-item"><%= link_to "Item Pools", admin_item_pools_path %></li>
-                  <li class="menu-item"><%= link_to "Reservable Items", admin_reservable_items_path %></li>
                   <li class="menu-item"><%= link_to "Reservations", admin_reservations_path %></li>
                   <li class="menu-item"><%= link_to "Pickups", admin_pickups_path %></li>
                 <% end %>
