+++ config/routes.rb
@@ -169,9 +169,9 @@ Rails.application.routes.draw do
       resources :item_pools do
         scope module: "item_pools" do
           resource :availability, only: :show
+          resources :reservable_items
         end
       end
-      resources :reservable_items
       resources :reservations do
         scope module: "reservations" do
           resource :review, only: [:edit, :update]
